+++ src/subHeader/SubHeader.style.tsx
@@ -1,15 +1,12 @@
 import styled from 'styled-components'
 
-import { font, space } from '../_utils/branding'
+import { space } from '../_utils/branding'
 import { normalizeHorizontally } from '../layout/layoutNormalizer'
-import { Title } from '../title'
+import { TextSubHeaderStrong } from '../typography/subHeaderStrong'
 
-export const StyledSubHeader = styled(Title)`
-  font-size: ${font.l.size};
-  line-height: ${font.l.lineHeight};
+export const StyledSubHeader = styled(TextSubHeaderStrong)`
+  ${normalizeHorizontally};
+
   padding-top: ${space.xl};
   padding-bottom: ${space.m};
-  margin: 0;
-
-  ${normalizeHorizontally};
 `