+++ src/subHeader/SubHeader.tsx
@@ -3,9 +3,10 @@ import React from 'react'
 import { StyledSubHeader } from './SubHeader.style'
 
 export type SubHeaderProps = Readonly<{
-  children: string
+  children: React.ReactNode
+  as?: keyof JSX.IntrinsicElements | React.ComponentType<any> // from styled-components types
 }>
 
-export const SubHeader = ({ children }: SubHeaderProps) => (
-  <StyledSubHeader headingLevel="2">{children}</StyledSubHeader>
+export const SubHeader = ({ children, as = 'h2' }: SubHeaderProps) => (
+  <StyledSubHeader as={as}>{children}</StyledSubHeader>
 )