+++ src/typography/Text.tsx
@@ -3,6 +3,7 @@ import React, { ReactNode } from 'react'
 import { replaceNewLineWithBR } from '../_utils'
 
 export type TextProps = Readonly<{
+  id?: string
   className?: string
   children: string | ReactNode
   /**
@@ -19,11 +20,11 @@ export type TextProps = Readonly<{
   itemProp?: string
 }>
 
-export const Text = ({ children, className, isInverted, isDisabled, ...props }: TextProps) => {
+export const Text = ({ children, id, className, isInverted, isDisabled, ...props }: TextProps) => {
   const content = typeof children === 'string' ? replaceNewLineWithBR(children) : children
 
   return (
-    <span className={className} {...props}>
+    <span id={id} className={className} {...props}>
       {content}
     </span>
   )