+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Unreleased
 
-**[UPDATE]** Allow polymorphism with `as` prop for `TheVoice`
+- **[UPDATE]** Allow polymorphism with `as` prop for `SubHeader`
+- **[UPDATE]** Allow polymorphism with `as` prop for `TheVoice`
 - [...]
 
 # v55.0.0 (18/03/2021)