+++ CHANGELOG.md
@@ -1,7 +1,19 @@
 # Unreleased
 
 - **[NEW]** New `CardsStackSection`
+- **[UPDATE]** Allow polymorphism with `as` prop for `SubHeader`
+- **[UPDATE]** Allow polymorphism with `as` prop for `TheVoice`
+- [...]
+
+# v55.0.0 (18/03/2021)
+
+- **[BREAKING CHANGE]** Remove root exports
+- **[UPDATE]** Allow newer node/npm version
+
+# v54.0.1 (12/03/2021)
+
 - **[FIX]** Fixed `ChevronIcon` direction on `SelectField`
+- **[FIX]** Export `ItemActionStatus`
 
 # v54.0.0 (09/03/2021)